# internal/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package database

import (
	"database/sql"

	"github.com/google/uuid"
)

type Song struct {
	ID          uuid.UUID
	GroupName   string
	SongName    string
	ReleaseDate sql.NullTime
	Text        sql.NullString
	Link        sql.NullString
}